[ci skip] refac expect/actual declaration to be in one place
