--- template/app/src/androidMain/kotlin/kotli/app/di/DI.android.kt
@@ -1,12 +1,0 @@
-package kotli.app.di
-
-import kotli.app.di.data.dataStoreSourceModule
-import kotli.app.di.data.roomSourceModule
-import org.koin.core.KoinApplication
-
-actual fun configure(app: KoinApplication) {
-    app.modules(
-        roomSourceModule,
-        dataStoreSourceModule
-    )
-}
