--- template/app/src/androidMain/kotlin/kotli/app/di/data/DataStoreSourceModule.android.kt
@@ -1,8 +1,0 @@
-package kotli.app.di.data
-
-import kotli.app.Application
-
-actual fun createDataStorePath(fileName: String): String {
-    val context = Application.instance
-    return context.filesDir.resolve(fileName).absolutePath
-}
