+++ template/app/src/androidMain/kotlin/kotli/app/factory/createDataStorePath.kt
@@ -1,0 +1,8 @@
+package kotli.app.factory
+
+import kotli.app.Application
+
+actual fun createDataStorePath(fileName: String): String {
+    val context = Application.instance
+    return context.filesDir.resolve(fileName).absolutePath
+}
