+++ template/app/src/commonMain/kotlin/kotli/app/di/presentation/AppModule.kt
@@ -25,7 +25,7 @@ import kotli.app.presentation.showcases.userflow.navigation.args.to.ArgsNavigati
 import kotli.app.presentation.showcases.userflow.navigation.no_args.from.NoArgsNavigationFromViewModel
 import kotli.app.presentation.showcases.userflow.navigation.no_args.to.NoArgsNavigationToViewModel
 import kotli.app.presentation.loader.LoaderViewModel
-import kotli.app.presentation.showcases.dataflow.room.crud.createRoomCrudViewModel
+import kotli.app.factory.createRoomCrudViewModel
 import kotli.app.presentation.template.screen_with_args.TemplateViewModel
 import kotli.app.presentation.template.screen_without_args.TemplateNoArgsViewModel
 import kotli.app.presentation.theme.AppThemePersistenceViewModel
