+++ template/app/src/commonMain/kotlin/kotli/app/factory/configureKoin.kt
@@ -1,0 +1,5 @@
+package kotli.app.factory
+
+import org.koin.core.KoinApplication
+
+expect fun configureKoin(app: KoinApplication)
