+++ template/app/src/commonMain/kotlin/kotli/app/factory/createRoomCrudViewModel.kt
@@ -1,0 +1,5 @@
+package kotli.app.factory
+
+import kotli.app.presentation.showcases.dataflow.room.crud.RoomCrudViewModel
+
+expect fun createRoomCrudViewModel(): RoomCrudViewModel
