+++ template/app/src/commonMain/kotlin/kotli/app/factory/createSqlDriver.kt
@@ -1,0 +1,5 @@
+package kotli.app.factory
+
+import app.cash.sqldelight.db.SqlDriver
+
+expect fun createSqlDriver(name: String): SqlDriver
