+++ template/app/src/commonMain/kotlin/kotli/app/presentation/showcases/dataflow/room/crud/RoomCrudViewModel.kt
@@ -5,8 +5,6 @@ import shared.presentation.navigation.NavigationStore
 import shared.presentation.store.DataState
 import shared.presentation.viewmodel.BaseViewModel
 
-expect fun createRoomCrudViewModel(): RoomCrudViewModel
-
 abstract class RoomCrudViewModel(val navigationStore: NavigationStore) : BaseViewModel() {
 
     val usersState = DataState<List<UserData>>(emptyList())
