--- template/app/src/iosMain/kotlin/kotli/app/data/source/database/sqldelight/AppSqlDelightSource.ios.kt
@@ -1,9 +1,0 @@
-package kotli.app.data.source.database.sqldelight
-
-import app.cash.sqldelight.async.coroutines.synchronous
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.native.NativeSqliteDriver
-
-actual fun createSqlDriver(name: String): SqlDriver {
-    return NativeSqliteDriver(AppDatabase.Schema.synchronous(), name)
-}
