--- template/app/src/iosMain/kotlin/kotli/app/di/data/DataStoreSourceModule.ios.kt
@@ -1,18 +1,0 @@
-package kotli.app.di.data
-
-import platform.Foundation.NSDocumentDirectory
-import platform.Foundation.NSFileManager
-import platform.Foundation.NSURL
-import platform.Foundation.NSUserDomainMask
-
-actual fun createDataStorePath(fileName: String): String {
-    val documentDirectory: NSURL? = NSFileManager.defaultManager.URLForDirectory(
-        directory = NSDocumentDirectory,
-        inDomain = NSUserDomainMask,
-        appropriateForURL = null,
-        create = false,
-        error = null,
-    )
-    val directory = requireNotNull(documentDirectory).path!!
-    return "${directory}/$fileName"
-}
