+++ template/app/src/iosMain/kotlin/kotli/app/factory/createDataStorePath.kt
@@ -1,0 +1,18 @@
+package kotli.app.factory
+
+import platform.Foundation.NSDocumentDirectory
+import platform.Foundation.NSFileManager
+import platform.Foundation.NSURL
+import platform.Foundation.NSUserDomainMask
+
+actual fun createDataStorePath(fileName: String): String {
+    val documentDirectory: NSURL? = NSFileManager.defaultManager.URLForDirectory(
+        directory = NSDocumentDirectory,
+        inDomain = NSUserDomainMask,
+        appropriateForURL = null,
+        create = false,
+        error = null,
+    )
+    val directory = requireNotNull(documentDirectory).path!!
+    return "${directory}/$fileName"
+}
