--- template/app/src/iosX64Main/kotlin/kotli/app/data/source/database/room/AppRoomSource.ios.kt
@@ -1,13 +1,0 @@
-package kotli.app.data.source.database.room
-
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import platform.Foundation.NSHomeDirectory
-
-actual fun getRoomDatabaseBuilder(name: String): RoomDatabase.Builder<AppDatabase> {
-    val dbFilePath = NSHomeDirectory() + "/$name"
-    return Room.databaseBuilder<AppDatabase>(
-        name = dbFilePath,
-        factory = { AppDatabase::class.instantiateImpl() }
-    )
-}
