--- template/app/src/jsMain/kotlin/kotli/app/di/DI.js.kt
@@ -1,6 +1,0 @@
-package kotli.app.di
-
-import org.koin.core.KoinApplication
-
-actual fun configure(app: KoinApplication) {
-}
