+++ template/app/src/jsMain/kotlin/kotli/app/factory/configureKoin.kt
@@ -1,0 +1,6 @@
+package kotli.app.factory
+
+import org.koin.core.KoinApplication
+
+actual fun configureKoin(app: KoinApplication) {
+}
