+++ template/app/src/jsMain/kotlin/kotli/app/factory/createSqlDriver.kt
@@ -1,0 +1,13 @@
+package kotli.app.factory
+
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.driver.worker.WebWorkerDriver
+import org.w3c.dom.Worker
+
+actual fun createSqlDriver(name: String): SqlDriver {
+    return WebWorkerDriver(
+        Worker(
+            js("""new URL("@cashapp/sqldelight-sqljs-worker/sqljs.worker.js", import.meta.url)""")
+        )
+    )
+}
