--- template/app/src/jvmMain/kotlin/kotli/app/data/source/database/sqldelight/AppSqlDelightSource.jvm.kt
@@ -1,8 +1,0 @@
-package kotli.app.data.source.database.sqldelight
-
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
-
-actual fun createSqlDriver(name: String): SqlDriver {
-    return JdbcSqliteDriver("jdbc:sqlite:$name")
-}
