+++ template/app/src/jvmMain/kotlin/kotli/app/factory/createDataStorePath.kt
@@ -1,0 +1,8 @@
+package kotli.app.factory
+
+import java.io.File
+
+actual fun createDataStorePath(fileName: String): String {
+    val file = File(System.getProperty("java.io.tmpdir"), fileName)
+    return file.absolutePath
+}
