+++ template/app/src/jvmMain/kotlin/kotli/app/factory/createSqlDriver.kt
@@ -1,0 +1,8 @@
+package kotli.app.factory
+
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
+
+actual fun createSqlDriver(name: String): SqlDriver {
+    return JdbcSqliteDriver("jdbc:sqlite:$name")
+}
