--- template/app/src/mobileAndDesktopMain/kotlin/kotli/app/di/data/RoomSourceModule.kt
@@ -1,8 +1,0 @@
-package kotli.app.di.data
-
-import kotli.app.data.source.database.room.AppRoomSource
-import org.koin.dsl.module
-
-val roomSourceModule = module {
-    single { AppRoomSource() }
-}
