+++ template/app/src/mobileAndDesktopMain/kotlin/kotli/app/factory/createDataStorePath.kt
@@ -1,0 +1,3 @@
+package kotli.app.factory
+
+expect fun createDataStorePath(fileName: String = "app.preferences_pb"): String
